stage4/captcha_generator/requirements.txt
flask
Pillow
gunicorn